changing counter to uint64

+++ common/rpc-models.go
@@ -335,7 +335,7 @@ type ListJobSummaryResponse struct {
 	SkippedSymlinkCount     uint32 `json:",string"`
 	HardlinksConvertedCount uint32 `json:",string"`
 	SkippedSpecialFileCount uint32 `json:",string"`
-	SkippedArchiveFileCount uint32 `json:",string"`
+	SkippedArchiveFileCount uint64 `json:",string"`
 }
 
 // wraps the standard ListJobSummaryResponse with sync-specific stats
